♻️ Extract components (#847)
